Move upload repo metho to AtomicStructureRepository
