+++ app/assets/javascripts/3dbio_viewer/package.json
@@ -45,7 +45,7 @@
         "build": "INLINE_RUNTIME_CHUNK=false react-scripts build && yarn create-build-symlinks",
         "create-build-symlinks": "ts-node -O '{\"module\":\"commonjs\"}' src/scripts/create-build-symlinks.ts",
         "lint": "eslint . --ext .ts,.tsx",
-        "prettify": "prettier \"./**/*.{js,jsx,json,css,ts,tsx}\" --write"
+        "prettify": "prettier \"./src/**/*.{js,jsx,json,css,ts,tsx}\" --write"
     },
     "eslintConfig": {
         "extends": [
