+++ app/assets/javascripts/3dbio_viewer/src/compositionRoot.ts
@@ -1,3 +1,4 @@
+import { BionotesAtomicStructureRepository } from "./data/repositories/BionotesAtomicStructureRepository";
 import { EbiDbModelRepository } from "./data/repositories/EbiDbModelRepository";
 import { ApiPdbRepository } from "./data/repositories/protvista/ApiPdbRepository";
 import { GetPdbUseCase } from "./domain/usecases/GetPdbUseCase";
@@ -7,11 +8,12 @@ import { UploadAtomicStructureUseCase } from "./domain/usecases/UploadAtomicStru
 export function getCompositionRoot() {
     const pdbRepository = new ApiPdbRepository();
     const dbModelRepository = new EbiDbModelRepository();
+    const atomicStructureRepository = new BionotesAtomicStructureRepository();
 
     return getExecute({
         getPdb: new GetPdbUseCase(pdbRepository),
         searchDbModels: new SearchDbModelsUseCase(dbModelRepository),
-        uploadAtomicStructure: new UploadAtomicStructureUseCase(dbModelRepository),
+        uploadAtomicStructure: new UploadAtomicStructureUseCase(atomicStructureRepository),
     });
 }
