+++ app/assets/javascripts/3dbio_viewer/src/domain/entities/AtomicStructureRepository.ts
@@ -1,0 +1,12 @@
+import { AtomicStructure } from "./AtomicStructure";
+import { FutureData } from "./FutureData";
+
+export interface AtomicStructureRepository {
+    build(options: BuildOptions): FutureData<AtomicStructure>;
+}
+
+export interface BuildOptions {
+    structureFile: File;
+    jobTitle?: string;
+    annotationsFile?: File;
+}
