+++ app/assets/javascripts/3dbio_viewer/src/domain/repositories/DbModelRepository.ts
@@ -1,19 +1,11 @@
 import { DbModelCollection, DbModelType } from "../entities/DbModel";
 import { FutureData } from "../entities/FutureData";
-import { AtomicStructure } from "../entities/AtomicStructure";
 
 export interface DbModelRepository {
     search(options: SearchOptions): FutureData<DbModelCollection>;
-    upload(options: UploadOptions): FutureData<AtomicStructure>;
 }
 
 export interface SearchOptions {
     query: string;
     type?: DbModelType;
-}
-
-export interface UploadOptions {
-    structureFile: File;
-    jobTitle?: string;
-    annotationsFile?: File;
 }
